Molly (#7)
* Add files via upload

I'm still working on the navigation behind everything but this is what I've done so far. The blank screen is for more questions that the person using will want to ask the patients.

* Add files via upload

This is the updated storyboard with navigation. I am still working on individual parts but you can navigate throughout the app now

* Add files via upload

* Add files via upload

Co-authored-by: mollylynch <65363138+mollylynch@users.noreply.github.com>